+++ src/oem_prov_storage.c
@@ -1,12 +1,14 @@
 // SPDX-License-Identifier: BSD-3-Clause
 /*
- * Copyright 2025 NXP
+ * Copyright 2025-2026 NXP
  */
 
 #include <smw_keymgr.h>
 
 #include "oem_prov_status.h"
 #include "oem_prov_debug_info.h"
+#include "oem_prov_os.h"
+#include "oem_prov_arithmetic_ops.h"
 
 int oem_prov_commit_key_storage(void)
 {
@@ -23,6 +25,16 @@ int oem_prov_commit_key_storage(void)
 
 	struct smw_commit_key_storage_args smw_args = { 0 };
 	int res = SMW_STATUS_OK;
+	static const char *const warnings[] = {
+		"WARNING: You are about to commit the non-volatile key storage.",
+		"This operation is IRREVERSIBLE and will:",
+		"  - Push the active key storage to physical memory",
+		"  - Increment the hardware anti-rollback counter"
+	};
+
+	status = oem_prov_confirm_operation(warnings, ARRAY_SIZE(warnings));
+	if (status != OEM_PROV_STATUS_OK)
+		return status;
 
 	smw_args.subsystem_name = SMW_SUBSYSTEM_NAME_ELE;
 	res = smw_commit_key_storage(&smw_args);