+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This file briefly describes the list of changes and bug fixes for each OEM Provi
 ### **Release 2.3**
 * Enhanced documentation
 * Retrieve the device lifecycle
+* Add user confirmation for irreversible operations
 
 ### **Release 2.2**
 * Updated the Yocto layer example to support whinlatter